+++ deploy/def/job.go
@@ -8,8 +8,8 @@ import (
 	"fmt"
 
 	"github.com/go-ozzo/ozzo-validation"
-	"github.com/hyperledger/burrow/deploy/abi"
 	"github.com/hyperledger/burrow/deploy/def/rule"
+	"github.com/hyperledger/burrow/execution/evm/abi"
 )
 
 //TODO: Interface all the jobs, determine if they should remain in definitions or get their own package
