--- execution/evm/abi/types.go
@@ -1,64 +1,0 @@
-// Copyright 2017 Monax Industries Limited
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//    http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package abi
-
-import "github.com/hyperledger/burrow/execution/evm/sha3"
-
-// Ethereum defines types and calling conventions for the ABI
-// (application binary interface) here: https://github.com/ethereum/wiki/wiki/Ethereum-Contract-ABI
-// We make a start of representing them here
-
-// We use TypeName rather than Type to reserve 'Type' for a possible future
-// ABI type the can hold an ABI-native type mapping
-type TypeName string
-
-type Arg struct {
-	Name     string
-	TypeName TypeName
-}
-
-type Return struct {
-	Name     string
-	TypeName TypeName
-}
-
-const (
-	// We don't need to be exhaustive here, just make what we used strongly typed
-	AddressTypeName TypeName = "address"
-	Uint64TypeName  TypeName = "uint64"
-	Bytes32TypeName TypeName = "bytes32"
-	BoolTypeName    TypeName = "bool"
-)
-
-const (
-	FunctionSelectorLength = 4
-	AddressLength          = 20
-)
-
-type FunctionSelector [FunctionSelectorLength]byte
-
-func (fs FunctionSelector) Bytes() []byte {
-	return fs[:]
-}
-
-func FunctionID(signature string) FunctionSelector {
-	return FirstFourBytes(sha3.Sha3([]byte(signature)))
-}
-
-func FirstFourBytes(byteSlice []byte) [4]byte {
-	var bs [4]byte
-	copy(bs[:], byteSlice[:4])
-	return bs
-}
